% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DownloadSenti4SD}
\alias{DownloadSenti4SD}
\title{Download Senti4sd}
\usage{
DownloadSenti4SD(version = "master", overwrite = FALSE)
}
\arguments{
\item{version}{Git version tag representing the version of
Senti4SD.}

\item{overwrite}{If FALSE, files that are already found in
RSenti4SD installation directory won't be downloaded.}
}
\description{
Downloads required Senti4SD files (LiblineaR model, jar file and
DSM) from GitHub.
}
