% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{ReadSenti4SDFeatures}
\alias{ReadSenti4SDFeatures}
\title{Read Features}
\usage{
ReadSenti4SDFeatures(filename)
}
\arguments{
\item{filename}{CSV file containing features.}
}
\value{
A data.table object with id column parsed as an integer.
}
\description{
Read feature file.
}
