% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{RunSenti4SD}
\alias{RunSenti4SD}
\title{Run Senti4SD}
\usage{
RunSenti4SD(
  df,
  model,
  chunk.size = 1000,
  memory.limit = 0,
  input.column = "text",
  prefix = NULL
)
}
\arguments{
\item{df}{A data.frame or data.table containing the}

\item{model}{The LiblineaR model to use for prediction.}

\item{chunk.size}{Maximum number of text element to consider for
one single run of Senti4SD.}

\item{memory.limit}{Maximum amount of memory (in GB) to use for one
run of Senti4SD. Overrides \code{chunk.size} by setting it to
\code{500 * memory.limit}.}

\item{input.column}{The name of the column in \code{df}
containing the text to analyze with Senti4SD.}

\item{prefix}{Optional prefix to add to the table's column names.}
}
\value{
A data.table object with columns from \code{df} and
  added polarity column.
}
\description{
Runs Senti4SD on a text column of a data.frame. Splits the input in
multiple chunks and running Senti4SD on each chunk.
}
