% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SD}
\alias{Senti4SD}
\title{Senti4SD}
\usage{
Senti4SD(
  text,
  model,
  chunk.size = 1000,
  memory.limit = 0,
  predict = TRUE,
  sparse.features = TRUE
)
}
\arguments{
\item{text}{A character vector on which to run Senti4SD.}

\item{model}{The LiblineaR model to use for prediction.}

\item{chunk.size}{Maximum number of text element to consider for
one single run of Senti4SD.}

\item{memory.limit}{Maximum amount of memory (in GB) to use for one
run of Senti4SD. Overrides \code{chunk.size} by setting it to
\code{500 * memory.limit}.}

\item{predict}{If TRUE makes prediction, otherwise return the
features.}

\item{sparse.features}{If predict is TRUE, returns a sparseMatrix
instead of a regular matrix.}
}
\value{
If predict is TRUE, a data.table object with text, id and
  (predicted) polarity columns, otherwise a matrix containing
  the features.
}
\description{
Runs Senti4SD on given pieces of text by splitting the input in
multiple chunks and running Senti4SD on each chunk.
}
