% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDFeatures}
\alias{Senti4SDFeatures}
\title{Features}
\usage{
Senti4SDFeatures(input, output, read.file = TRUE, use.temp.file = FALSE)
}
\arguments{
\item{input}{Input text file with a single column and no header.}

\item{output}{Output file containing computed features.}

\item{read.file}{If TRUE, read with data.table the content of
output file.}

\item{use.temp.file}{If TRUE, output will be removed after feature
computation. This also enforces read.file as TRUE.}
}
\value{
The result of feature extraction as data.table if read.file
  is TRUE, stdout and stderr of the Java process as invisible
  object otherwise.
}
\description{
Call Senti4SD-fast.jar to compute features.
}
