% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDPredict}
\alias{Senti4SDPredict}
\title{Predict}
\usage{
Senti4SDPredict(model, features)
}
\arguments{
\item{model}{The LiblineaR model object.}

\item{features}{The feature data.table object.}
}
\value{
A factor with levels positive, negative and neutral.
}
\description{
Predict polarity from a set of features.
}
