% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti4sd.R
\name{Senti4SDSingle}
\alias{Senti4SDSingle}
\title{Senti4SD Single}
\usage{
Senti4SDSingle(
  text,
  model,
  predict = TRUE,
  sparse.features = TRUE,
  tmpdir = getOption("senti4sd.tmpdir", tempdir())
)
}
\arguments{
\item{text}{A character vector on which to run Senti4SD.}

\item{model}{The LiblineaR model to use for prediction.}

\item{predict}{If TRUE makes prediction, otherwise return the
features.}

\item{sparse.features}{If predict is TRUE, returns a sparseMatrix
instead of a regular matrix.}

\item{tmpdir}{Directory where to store temporary files.}
}
\value{
If predict is TRUE, A data.table object with text, id and
  (predicted) polarity columns, otherwise a matrix containing the
  features.
}
\description{
Runs Senti4SD on given pieces of text.
}
